+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### Changes
 - Added trufflehog to pre-commit ([#1441](https://github.com/neptune-ai/neptune-client/pull/1441))
+- Identify client's artifact supported version by adding `X-Neptune-Artifact-Api-Version` header to get artifact attribute request ([#1436](https://github.com/neptune-ai/neptune-client/pull/1436))
 
 ## neptune 1.6.1